requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
pyjwt
werkzeug
jinja2
python-dotenv   # if using .env for EMAIL creds
